Use a catch-all target and track pending drops
The skeeball game is built such that there is no dedicated hole for the
missed target goal and all balls will pass through this area. As a
result, the sensor will need to be placed where all balls will pass by
it, even balls that have already hit a target.

To resolve this and still give the feeling of instant feedback when the
game is over, a pending_drops counter is added. This keeps track of the
number of balls 'in flight' between the scoring targets and the
catch-all area.